+++ src/main/java/net/spell_engine/api/item/trinket/SpellBookTrinketItem.java
@@ -1,6 +1,9 @@
 package net.spell_engine.api.item.trinket;
 
+import dev.emi.trinkets.api.SlotReference;
 import dev.emi.trinkets.api.TrinketItem;
+import net.minecraft.entity.LivingEntity;
+import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.Identifier;
 
@@ -22,4 +25,12 @@ public class SpellBookTrinketItem extends TrinketItem implements SpellBookItem {
         return poolId;
     }
 
+    @Override
+    public boolean canUnequip(ItemStack stack, SlotReference slot, LivingEntity entity) {
+        var isOnCooldown = false;
+        if (entity instanceof PlayerEntity player) {
+            isOnCooldown = player.getItemCooldownManager().isCoolingDown(stack.getItem());
+        }
+        return super.canUnequip(stack, slot, entity) && !isOnCooldown;
+    }
 }